requirements.txt
gunicorn
gevent
gevent-websocket
flask
flask_socketio
yt_dlp
pyyaml